+++ packages/web/src/App.tsx
@@ -1,0 +1,46 @@
+import { useState, useTransition } from 'react'
+import reactLogo from './assets/react.svg'
+import viteLogo from '/vite.svg'
+
+function App() {
+  const [count, setCount] = useState(0)
+  const [isPending, startTransition] = useTransition()
+
+  const handleClick = () => {
+    // Urgent Update
+    setCount(count => count + 1)
+
+    // Start a non-urgent update
+    startTransition(() => {
+      // Simulating a heavy computation
+      const startTime = Date.now()
+      while (Date.now() - startTime < 100) {
+        // Artificial delay
+      }
+      setCount(count => count + 1)
+    })
+  }
+
+  return (
+    <div className={'bg-light-optiblue11'}>
+      <div>
+        <a href="https://vitejs.dev" target="_blank">
+          <img src={viteLogo} className="logo" alt="Vite logo" />
+        </a>
+        <a href="https://react.dev" target="_blank">
+          <img src={reactLogo} className="logo react" alt="React logo" />
+        </a>
+      </div>
+      <h1>Vite + React</h1>
+      <div className="card">
+        <button onClick={() => setCount(count => count + 1)}>count is {count}</button>
+        <p>
+          Edit <code>src/App.jsx</code> and save to test HMR
+        </p>
+      </div>
+      <p className="read-the-docs">Click on the Vite and React logos to learn more</p>
+    </div>
+  )
+}
+
+export default App