+++ shared/CHANGELOG.md
@@ -3,22 +3,74 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## 1.0.23 (2024-09-20)
+## 1.0.18 (2024-09-19)
 
 **Note:** Version bump only for package afrikit-shared
 
-## 1.0.22 (2024-09-20)
+## [1.0.17](https://github.com/AfrinvestOptimus/afrikit/compare/afrikit-shared@1.0.15...afrikit-shared@1.0.17) (2024-09-19)
 
 **Note:** Version bump only for package afrikit-shared
 
-## 1.0.21 (2024-09-20)
+## 1.0.16 (2024-09-19)
 
 **Note:** Version bump only for package afrikit-shared
 
-## [1.0.20](https://github.com/AfrinvestOptimus/afrikit/compare/afrikit-shared@0.2.1...afrikit-shared@1.0.20) (2024-09-20)
+## 1.0.15 (2024-09-19)
 
 **Note:** Version bump only for package afrikit-shared
 
-## 0.2.1 (2024-09-20)
+## 1.0.14 (2024-09-19)
 
 **Note:** Version bump only for package afrikit-shared
+
+## 1.0.13 (2024-09-19)
+
+**Note:** Version bump only for package afrikit-shared
+
+## 1.0.12 (2024-09-19)
+
+**Note:** Version bump only for package afrikit-shared
+
+## 1.0.11 (2024-09-19)
+
+**Note:** Version bump only for package afrikit-shared
+
+## 1.0.10 (2024-09-19)
+
+**Note:** Version bump only for package afrikit-shared
+
+## 1.0.9 (2024-09-19)
+
+**Note:** Version bump only for package afrikit-shared
+
+## 1.0.8 (2024-09-19)
+
+**Note:** Version bump only for package afrikit-shared
+
+## 1.0.7 (2024-09-19)
+
+**Note:** Version bump only for package afrikit-shared
+
+## 1.0.6 (2024-09-19)
+
+**Note:** Version bump only for package afrikit-shared
+
+## 1.0.5 (2024-09-19)
+
+**Note:** Version bump only for package afrikit-shared
+
+## 1.0.4 (2024-09-19)
+
+**Note:** Version bump only for package afrikit-shared
+
+## 1.0.3 (2024-09-18)
+
+**Note:** Version bump only for package afrikit-shared
+
+## 1.0.2 (2024-09-18)
+
+**Note:** Version bump only for package afrikit-shared
+
+## 1.0.1 (2024-09-18)
+
+**Note:** Version bump only for package shared